+++ src/ETL-SQL.Connectors/Json/JsonConnector.cs
@@ -24,6 +24,7 @@ namespace ETL_SQL.Connectors.Json
 
         public Dictionary<string, string[]> GetSupportedOptions() => new(StringComparer.OrdinalIgnoreCase)
         {
+            { "PATH", Array.Empty<string>() },
             { "ROOT_PATH", Array.Empty<string>() },
             { "COMPRESS", new[] { "ON", "OFF" } },
             { "ENCRYPT", new[] { "ON", "OFF" } },